% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantTxDbHub-package.R
\docType{package}
\name{PlantTxDbHub-package}
\alias{PlantTxDbHub-package}
\alias{PlantTxDbHub}
\title{PlantTxDbHub: AnnotationHub‑based TxDb databases for
plants}
\description{
Provides TxDb (transcript database) annotations for plant species, initially including Arabidopsis thaliana (TAIR10), Oryza sativa (IRGSP-1.0), and Glycine max (Wm82 v2.1), based on Ensembl Plants release 62. The SQLite files are hosted on Zenodo (or other permanent URLs) and are downloaded on demand via the `downloadPlantTxDbs()` function. The list of available species and their download URLs is maintained in a curated CSV file (inst/extdata/metadata.csv), allowing users and contributors to easily add new databases without modifying R code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kabilanbio/PlantTxDbHub}
  \item Report bugs at \url{https://github.com/kabilanbio/PlantTxDbHub/issues}
}

}
\author{
Kabilan S
}
\keyword{package}
